fixed linker error

+++ soccer/src/soccer/strategy/coordinator/marking.hpp
@@ -37,10 +37,10 @@ private:
     static constexpr double kSuperDangerSub = 3.1415926535;
     int num_markers_;
 
-    std::array<int, kNumShells> marking_list_{};  // Initialize it to invalid robot id in constructor
-    std::array<int, kNumShells> danger_score_{}; // infinity initialized in constructor, no one is a valid target initially
-    std::array<int, kNumShells> enemey_to_friends_{};
-    std::vector<int> queue_;
+    std::array<uint8_t, kNumShells> marking_list_{};  // Initialize it to invalid robot id in constructor
+    std::array<double, kNumShells> danger_score_{}; // infinity initialized in constructor, no one is a valid target initially
+    std::array<uint8_t, kNumShells> enemey_to_friends_{};
+    std::vector<uint8_t> queue_;
     WorldState last_world_state_;
     FieldDimensions field_dimensions_ = FieldDimensions::kDefaultDimensions;
     rclcpp::Subscription<rj_msgs::msg::WorldState>::SharedPtr world_state_sub_;
